+++ src/test/java/no/nav/tag/tiltaksgjennomforing/avtale/TestData.java
@@ -471,13 +471,13 @@ public class TestData {
         endreAvtale.setTilrettelegging("Ingen");
         endreAvtale.setStartDato(Now.localDate());
         endreAvtale.setSluttDato(endreAvtale.getStartDato().plusMonths(12).minusDays(1));
-        endreAvtale.setStillingprosent(50.0);
+        endreAvtale.setStillingprosent(BigDecimal.valueOf(50.0));
         endreAvtale.setArbeidsoppgaver("Butikkarbeid");
         endreAvtale.setArbeidsgiverKontonummer("000111222");
         endreAvtale.setStillingstittel("Butikkbetjent");
         endreAvtale.setManedslonn(6808);
         endreAvtale.setStillingstype(Stillingstype.FAST);
-        endreAvtale.setAntallDagerPerUke(5.0);
+        endreAvtale.setAntallDagerPerUke(BigDecimal.valueOf(5.0));
         endreAvtale.setHarFamilietilknytning(false);
         return endreAvtale;
     }
@@ -577,8 +577,8 @@ public class TestData {
         endreAvtale.setStillingStyrk08(5223);
         endreAvtale.setStillingKonseptId(112968);
         endreAvtale.setArbeidsoppgaver("Butikkarbeid");
-        endreAvtale.setStillingprosent(50.5);
-        endreAvtale.setAntallDagerPerUke(5.0);
+        endreAvtale.setStillingprosent(BigDecimal.valueOf(50.5));
+        endreAvtale.setAntallDagerPerUke(BigDecimal.valueOf(5.0));
         endreAvtale.getMaal().add(TestData.etMaal());
         endreAvtale.setStartDato(Now.localDate());
         endreAvtale.setSluttDato(endreAvtale.getStartDato().plusMonths(12).minusDays(1));
@@ -605,7 +605,7 @@ public class TestData {
         endreAvtale.setTilrettelegging("Ingen");
         endreAvtale.setStartDato(Now.localDate());
         endreAvtale.setSluttDato(endreAvtale.getStartDato().plusMonths(12).minusDays(1));
-        endreAvtale.setStillingprosent(50.7);
+        endreAvtale.setStillingprosent(BigDecimal.valueOf(50.7));
         endreAvtale.setArbeidsoppgaver("Butikkarbeid");
         endreAvtale.setArbeidsgiverKontonummer("000111222");
         endreAvtale.setStillingstittel("Butikkbetjent");
@@ -617,7 +617,7 @@ public class TestData {
         endreAvtale.setArbeidsgiveravgift(BigDecimal.ONE);
         endreAvtale.setOtpSats(0.02);
         endreAvtale.setStillingstype(Stillingstype.FAST);
-        endreAvtale.setAntallDagerPerUke(5.0);
+        endreAvtale.setAntallDagerPerUke(BigDecimal.valueOf(5.0));
         endreAvtale.setRefusjonKontaktperson(new RefusjonKontaktperson("Ola", "Olsen", "12345678", true));
         return endreAvtale;
     }
